デビュー戦! ⚡️🏃🏻💨
MWC. Internship about to start. Looking for new projects to collaborate with. Code development forced to be done at full speed. The way of uploading it to Github is still not optimal. JS I need to learn how to do it without it all being in the script of a page. There is still a lot to improve in Backend with Python. Let's see if I can develop as much as I'd like in the little time I have left. がんばって!